Added command line option to tag unknown words in output (#48)
A new command line option --replace_unk_tagged has been added to
mirror the functionality in OpenNMT
(http://opennmt.net/OpenNMT/translation/unknowns/).
The option works just like --replace_unk, except that it wraps
the unknown word in a ｟unk:xxxxx｠tag. This is useful for
post-processing the unknown tokens in the output.